refactor: split MainLayout responsibilities into focused custom hooks
Extract filter state/computations into useTransactionFilters, CRUD/export
handlers into useTransactionHandlers, and modal open/close state into
useModalState. MainLayout is now JSX wiring only. Zero behaviour change —
all 86 test suites pass, tsc clean, build passes.

Closes #65
